Consolidate eventos visual tests
